+++ wds/cem-analyzer-plugin.js
@@ -5,6 +5,14 @@ const path = '/dist/custom-elements.json';
 // This plugin generates and serves the CEM on demand.
 export const cemAnalyzerPlugin = {
   name: 'cem-analyzer-plugin',
+  // we disable transform cache to enforce generation for every call
+  transform (context) {
+    if (context.path === path) {
+      return {
+        transformCache: false,
+      };
+    }
+  },
   async serve (context) {
     if (context.path === path) {
       const cemJson = await cli({